+++ packages/frontend/vite.config.ts
@@ -1,7 +1,7 @@
 import path from 'path';
 import pluginReplace from '@rollup/plugin-replace';
 import pluginVue from '@vitejs/plugin-vue';
-import { type UserConfig, defineConfig } from 'vite';
+import {type UserConfig, defineConfig} from 'vite';
 // @ts-expect-error https://github.com/sxzz/unplugin-vue-macros/issues/257#issuecomment-1410752890
 import ReactivityTransform from '@vue-macros/reactivity-transform/vite';
 
@@ -9,6 +9,7 @@ import locales from '../../locales';
 import meta from '../../package.json';
 import pluginUnwindCssModuleClassName from './lib/rollup-plugin-unwind-css-module-class-name';
 import pluginJson5 from './vite.json5';
+import compression from "vite-plugin-compression2";
 
 const extensions = ['.ts', '.tsx', '.js', '.jsx', '.mjs', '.json', '.json5', '.svg', '.sass', '.scss', '.css', '.vue'];
 
@@ -44,12 +45,12 @@ function toBase62(n: number): string {
 export function getConfig(): UserConfig {
 	return {
 		base: '/vite/',
-
 		server: {
 			port: 5173,
 		},
 
 		plugins: [
+			compression({ algorithm: 'brotliCompress'}),
 			pluginVue({
 				reactivityTransform: true,
 			}),